requirements.txt
django
django-cors-headers
djangorestframework
django-db-connection-pool[mysql]
